+++ build_modules/lib/build_modules.dart
@@ -2,7 +2,8 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-export 'src/errors.dart' show MissingModulesException;
+export 'src/errors.dart'
+    show MissingModulesException, UnsupportedModulesException;
 export 'src/kernel_builder.dart' show KernelBuilder, multiRootScheme;
 export 'src/meta_module_builder.dart'
     show MetaModuleBuilder, metaModuleExtension;