+++ build_web_compilers/lib/src/dev_compiler_bootstrap.dart
@@ -29,7 +29,24 @@ Future<void> bootstrapDdc(BuildStep buildStep) async {
       .decode(await buildStep.readAsString(moduleId)) as Map<String, dynamic>);
 
   // First, ensure all transitive modules are built.
-  var transitiveDeps = await _ensureTransitiveModules(module, buildStep);
+  List<Module> transitiveDeps;
+  try {
+    transitiveDeps = await _ensureTransitiveModules(module, buildStep);
+  } on UnsupportedModulesException catch (e) {
+    var librariesString = (await e.exactLibraries(buildStep).toList())
+        .map((lib) => AssetId(lib.id.package,
+            lib.id.path.replaceFirst(moduleLibraryExtension, '.dart')))
+        .join('\n');
+    log.warning('''
+Skipping compiling ${buildStep.inputId} with ddc because some of its
+transitive libraries have sdk dependencies that not supported on this platform:
+
+$librariesString
+
+https://github.com/dart-lang/build/blob/master/docs/faq.md#how-can-i-resolve-skipped-compiling-warnings
+''');
+    return;
+  }
   var jsId = module.primarySource.changeExtension(jsModuleExtension);
   var appModuleName = ddcModuleName(jsId);
   var appDigestsOutput =
@@ -108,10 +125,14 @@ Future<void> bootstrapDdc(BuildStep buildStep) async {
 final _lazyBuildPool = Pool(16);
 
 /// Ensures that all transitive js modules for [module] are available and built.
+///
+/// Throws an [UnsupportedModules] exception if there are any unsupported
+/// modules.
 Future<List<Module>> _ensureTransitiveModules(
     Module module, AssetReader reader) async {
   // Collect all the modules this module depends on, plus this module.
-  var transitiveDeps = await module.computeTransitiveDependencies(reader);
+  var transitiveDeps = await module.computeTransitiveDependencies(reader,
+      throwIfUnsupported: true);
   var jsModules = transitiveDeps
       .map((module) => module.primarySource.changeExtension(jsModuleExtension))
       .toList()