+++ build_vm_compilers/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - Support build_modules 2.0.0
   - Define our own `vm` platform and builders explicitly.
+- Skip trying to compile apps that import known incompatible libraries.
 
 ## 0.1.2